Fix var's names.

+++ schemedialog.cpp
@@ -21,20 +21,20 @@ void SchemeDialog::setPtr(std::shared_ptr<SchemeParams> &ptr)
 
 void SchemeDialog::on_OK_clicked()
 {
-    s_par_ptr->node_count = ui->node_count->value();
-    s_par_ptr->resistors = ui->resistors->value();
-    s_par_ptr->capasitors = ui->capacitors->value();
-    s_par_ptr->inductances = ui->inductances->value();
-    s_par_ptr->itun = ui->itun->value();
-    s_par_ptr->inun = ui->inun->value();
-    s_par_ptr->itut = ui->itut->value();
-    s_par_ptr->inut = ui->inut->value();
-    s_par_ptr->b_p_transistors = ui->bp_transistors->value();
-    s_par_ptr->u_p_transistors = ui->up_transistors->value();
-    s_par_ptr->oper_ampfilers = ui->amplifiers->value();
-    s_par_ptr->transformers = ui->transformers->value();
-    s_par_ptr->p_o_ampfilers = ui->p_o_alplifiers->value();
-    s_par_ptr->perf_transistors = ui->perf_tranformers->value();
+    s_par_ptr->node_count = static_cast<size_t>(ui->node_count->value());
+    s_par_ptr->resistors = static_cast<size_t>(ui->resistors->value());
+    s_par_ptr->capasitors = static_cast<size_t>(ui->capacitors->value());
+    s_par_ptr->inductances = static_cast<size_t>(ui->inductances->value());
+    s_par_ptr->itun = static_cast<size_t>(ui->itun->value());
+    s_par_ptr->inun = static_cast<size_t>(ui->inun->value());
+    s_par_ptr->itut = static_cast<size_t>(ui->itut->value());
+    s_par_ptr->inut = static_cast<size_t>(ui->inut->value());
+    s_par_ptr->b_p_transistors = static_cast<size_t>(ui->bp_transistors->value());
+    s_par_ptr->u_p_transistors = static_cast<size_t>(ui->up_transistors->value());
+    s_par_ptr->oper_ampfilers = static_cast<size_t>(ui->amplifiers->value());
+    s_par_ptr->transformers = static_cast<size_t>(ui->transformers->value());
+    s_par_ptr->p_o_ampfilers = static_cast<size_t>(ui->p_o_alplifiers->value());
+    s_par_ptr->perf_transistors = static_cast<size_t>(ui->perf_tranformers->value());
     this->hide();
     emit showMainWindow();
 }
